add coin transfer

+++ db/migrations/20210608203023_create_wallets_table.ts
@@ -4,7 +4,7 @@ export async function up(knex: Knex): Promise<void> {
     return knex.schema.createTable('wallets', (table) => {
         table.increments();
         table.string('discord_id').notNullable().unique();
-        table.integer('coins').notNullable();
+        table.float('coins').notNullable();
     });
 }
